fix search org. unit selection in user table
