+++ src/data/models/DHIS2Model.ts
@@ -1,4 +1,5 @@
-import { Id, NamedRef, Ref } from "../../domain/entities/Ref";
+import { GetType } from "purify-ts";
+import { NamedRef, Ref } from "../../domain/entities/Ref";
 import { Codec, Schema } from "../../utils/codec";
 
 export const RefModel: Codec<Ref> = Schema.object({
@@ -10,14 +11,10 @@ export const NamedRefModel: Codec<NamedRef> = Schema.object({
     name: Schema.optionalSafe(Schema.string, "Unknown"),
 });
 
-export type ApiD2OrgUnit = {
-    id: Id;
-    name: string;
-    path: string;
-};
-
-export const OrgUnitModel: Codec<ApiD2OrgUnit> = Schema.object({
+export const OrgUnitModel = Schema.object({
     id: Schema.string,
     name: Schema.string,
     path: Schema.string,
 });
+
+export type ApiD2OrgUnit = GetType<typeof OrgUnitModel>;
